// src/enums/privilege.ts
export enum Privilege {
  ChangeSiteSettings = 'ChangeSiteSettings',
  CompanyConnectionFlow = 'CompanyConnectionFlow',
  CompanyConnectionFlowReport = 'CompanyConnectionFlowReport',
  Connections = 'Connections',
  Contacts = 'Contacts',
  CuratorFiles = 'CuratorFiles',
  ElectronicDocument = 'ElectronicDocument',
  Equip = 'Equip',
  EquipMeasurements = 'EquipMeasurements',
  Mutual = 'Mutual',
  Order = 'Order',
  Payment = 'Payment',
  PlanChangeByConnection = 'PlanChangeByConnection',
  PlanChangeByConnectionCreate = 'PlanChangeByConnectionCreate',
  ScannedDocument = 'ScannedDocument',
  ScannedPassportGas = 'ScannedPassportGas',
  StatisticsByConnection = 'StatisticsByConnection',
  Telemetry = 'Telemetry',
  UnevennessByConnectionCreate = 'UnevennessByConnectionCreate',
  UnevennessPlanByConnection = 'UnevennessPlanByConnection',
  UserRoleManagement = 'UserRoleManagement',
  PartnerList = 'PartnerList',
}